# Draw power tracks faster by skipping samples that would be displayed on the same pixel