… src/test/components/__snapshots__/TrackPower.test.js.snap