When loading a large power profile, a lot of time is spent drawing the canvas of the power tracks. I think we can speed this up by avoiding the call to the canvas APIs when we would be drawing multiple times in a row at the same position.

Here are the profiles I've used to test:
- https://share.firefox.dev/45BHfJp 85k samples
- https://share.firefox.dev/3MYvCEO 785k samples (from https://bugzilla.mozilla.org/show_bug.cgi?id=1822483#c3)

And profiles of the profiler loading them multiple times first without and then with the patch:
- https://share.firefox.dev/3BZLReI - without the patch there's about 400ms of canvas work, with the patch it's instantaneous (ie. fast enough that the 1ms sampling rate is not enough to say for sure).
- https://share.firefox.dev/43qt1ck - the canvas work takes about 2.6s without the patch, 610ms with the patch. Still slower than I would like, but a lot more usable (less than a quarter of the total profile load time).

The shapes of the charts are almost the same with and without the patch (some lines are a little bit thicker without the patch). We could speed things up further if we accepted altering the shape a bit more.