@fqueze 2023-05-29
I followed [Julien's idea](https://github.com/firefox-devtools/profiler/pull/4637#issuecomment-1565227843) of showing only the min and max values. Here is a profile of repeatedly loading the 785k samples power profile: https://share.firefox.dev/3OGL1ef The work related to showing the canvases of power tracks is down to about 100ms, and the chart looks exactly like after #4634.